web/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-mpd2==3.1.1
httpx==0.28.1
